+++ .eslintrc.cjs
@@ -1,16 +1,8 @@
 module.exports = {
   root: true,
-  extends: ['@putro/typescript'],
+  extends: ['@putstack/typescript'],
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: ['./tsconfig.eslint.json'],
   },
-  overrides: [
-    {
-      files: ['packages/eslint-config-typescript/**/index.js', 'packages/prettier/index.js'],
-      rules: {
-        'unicorn/prefer-module': 'off',
-      },
-    },
-  ],
 };