+++ packages/eslint-config-typescript/tests/eslint-rule-tester.ts
@@ -1,0 +1,4 @@
+import { RuleTester } from '@typescript-eslint/rule-tester/dist';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const ruleTester = new RuleTester();