+++ packages/eslint-config-typescript/CHANGELOG.md
@@ -6,6 +6,6 @@ All notable changes to this configuration will be documented in this file.
 
 ### Features
 
-- First release of `@putro/eslint-config-typescript` that provides sensible, but strict, defaults for new and existing projects.
-  - If the default is too strict, there is also a `@putro/eslint-config-typescript/base` that is less strict that you can start out with.
+- First release of `@putstack/eslint-config-typescript` that provides sensible, but strict, defaults for new and existing projects.
+  - If the default is too strict, there is also a `@putstack/eslint-config-typescript/base` that is less strict that you can start out with.
   -